Rename dbSyner -> sync
